% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_functions.R
\name{.get_deseq_results}
\alias{.get_deseq_results}
\title{Function dispatched by getDESeqResults}
\usage{
.get_deseq_results(
  dds,
  contrast.numer,
  contrast.denom,
  sizeFactors,
  alpha,
  args.DESeq,
  args.results,
  ncores,
  quiet
)
}
\description{
Function dispatched by getDESeqResults
}
