% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_functions.R
\name{getDESeqDataSet}
\alias{getDESeqDataSet}
\title{Get DESeqDataSet objects for downstream analysis}
\usage{
getDESeqDataSet(
  dataset.list,
  regions.gr,
  sample_names = names(dataset.list),
  gene_names = NULL,
  sizeFactors = NULL,
  field = "score",
  ncores = detectCores(),
  quiet = FALSE
)
}
\arguments{
\item{dataset.list}{A list of GRanges datasets that can be individually
passed to \code{\link[BRGenomics:getCountsByRegions]{getCountsByRegions}}.}

\item{regions.gr}{A GRanges object containing regions of interest.}

\item{sample_names}{Names for each dataset in \code{dataset.list} are
required, and by default the names of the list elements are used. The names
must each contain the string "_rep#", where "#" is a single character
(usually a number) indicating the replicate. Sample names across different
replicates must be otherwise identical.}

\item{gene_names}{An optional character vector giving gene names, or any
other identifier over which reads should be counted. Gene names are
required if counting is to be performed over non-contiguous ranges, i.e. if
any genes have multiple ranges. If supplied, gene names are added to the
resulting \code{DESeqDataSet} object.}

\item{sizeFactors}{DESeq2 \code{sizeFactors} can be optionally applied in to
the \code{DESeqDataSet} object in this function, or they can be applied
later on, either by the user or in a call to \code{getDESeqResults}.
Applying the \code{sizeFactors} later is useful if multiple sets of factors
will be explored, although \code{sizeFactors} can be overwritten at any
time.}

\item{field}{Argument passed to \code{getCountsByRegions}.}

\item{ncores}{Number of cores to use for read counting across all samples.
Default is the total number of cores available.}

\item{quiet}{If \code{TRUE}, all output messages from call to
\code{\link[DESeq2:DESeqDataSet]{DESeqDataSet}} will be suppressed.}
}
\value{
A \code{DESeqData} object in which \code{rowData} are given as
  \code{rowRanges}, which are equivalent to \code{regions.gr}, unless there
  are non-contiguous gene regions (see note below). Samples (as seen in
  \code{colData}) are factored so that samples are grouped by
  \code{replicate} and \code{condition}, i.e. all non-replicate samples are
  treated as distinct, and the DESeq2 design = \code{~condition}.
}
\description{
This is a convenience function for generating \code{DESeqDataSet} objects,
but this function also adds support for counting reads across non-contiguous
regions.
}
\section{Use of non-contiguous gene regions}{
 In DESeq2, genes must be defined
  by single, contiguous chromosomal locations. This function allows
  individual genes to be encompassed by multiple distinct ranges in
  \code{regions.gr}. To use non-contiguous gene regions, provide
  \code{gene_names} in which some names are duplicated. For each unique gene
  in \code{gene_names}, this function will generate counts across all ranges
  for that gene, but be aware that it will only keep the largest range for
  each gene in the resulting \code{DESeqDataSet} object's \code{rowRanges}.
}

\section{A note on DESeq2 sizeFactors}{
 DESeq2 \code{sizeFactors} are
  sample-specific normalization that are applied by division, i.e.
  \eqn{counts_{norm,i}=counts_i / sizeFactor_i}{normcounts_i = counts_i /
  sizeFactor_i}. This is in contrast to normalization factors as defined in
  this package (and commonly elsewhere), which are applied by multiplication.
  Also note that DESeq2's "\code{normalizationFactors}" are not sample
  specific, but rather gene specific factors used to correct for
  ascertainment bias across different genes (e.g. as might be relevant for
  GSEA or Go analysis).
}

\seealso{
\code{\link[DESeq2:DESeqDataSet]{DESeqDataSet}},
  \code{\link[BRGenomics:getDESeqResults]{getDESeqResults}},
  \code{\link[BRGenomics:getDESeqResultsInBatch]{getDESeqResultsInBatch}}
}
\author{
Mike DeBerardine
}
