% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq_functions.R
\name{getDESeqResultsInBatch}
\alias{getDESeqResultsInBatch}
\title{Automate batch calls to getDESeqResults}
\usage{
getDESeqResultsInBatch(
  dds,
  sizeFactors = NULL,
  alpha = 0.05,
  anchor = NULL,
  permutations = FALSE,
  additional_comparisons = NULL,
  ncores = detectCores()
)
}
\arguments{
\item{ncores}{}
}
\value{

}
\description{
This function can automate the generation numerous pairwise DESeq2
comparisons using several logical schemes.
}
