% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikein_functions.R
\name{getSpikeInCounts}
\alias{getSpikeInCounts}
\alias{removeSpikeInReads}
\alias{getSpikeInReads}
\title{Filtering and counting spike-in reads}
\usage{
getSpikeInCounts(
  dataset.gr,
  si_pattern = NULL,
  si_names = NULL,
  field = "score",
  ncores = detectCores()
)

removeSpikeInReads(
  dataset.gr,
  si_pattern = NULL,
  si_names = NULL,
  field = "score",
  ncores = detectCores()
)

getSpikeInReads(
  dataset.gr,
  si_pattern = NULL,
  si_names = NULL,
  field = "score",
  ncores = detectCores()
)
}
\arguments{
\item{dataset.gr}{A GRanges object or a list of GRanges objects.}

\item{si_pattern}{A regular expression that matches spike-in chromosomes. Can
be used in addition to, or as an alternative to \code{si_names}.}

\item{si_names}{A character vector giving the names of the spike-in
chromosomes. Can be used in addition to, or as an alternative to
\code{si_pattern}.}

\item{field}{The metadata field in \code{dataset.gr} that contains
readcounts. If each range is an individual read, set \code{field = NULL}.}

\item{ncores}{The number of cores to use for computations.}
}
\value{
A dataframe containing total readcounts, experimental (non-spike-in)
  readcounts, and spike-in readcounts.
}
\description{
Filtering and counting spike-in reads
}
\author{
Mike DeBerardine
}
