% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import.bw_trim}
\alias{import.bw_trim}
\title{Import bigWig files (general)}
\usage{
import.bw_trim(
  file,
  genome = NULL,
  keep_X = TRUE,
  keep_Y = TRUE,
  keep_M = FALSE,
  keep_nonstandard = FALSE
)
}
\arguments{
\item{file}{Path of a bigWig file (non-stranded).}

\item{genome}{Optional string for UCSC reference genome, e.g. "hg38". If
given, non-standard chromosomes are trimmed.}

\item{keep_X}{Logical indicating whether the X chromosome should be kept.}

\item{keep_Y}{Logical indicating whether the Y chromosome should be kept.}

\item{keep_M}{Logical indicating whether mitochondrial chromosomes should be
kept.}
}
\value{
Imports a GRanges object
}
\description{
General function for importing a single bigWig file as a GRanges object. The
added functionality over \code{rtracklayer::import.bw} is in trimming odd
chromosomes.
}
\author{
Mike DeBerardine
}
