% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import.CoPRO}
\alias{import.CoPRO}
\title{Import CoPRO (or similar) bedGraph files}
\usage{
import.CoPRO(
  plus_file,
  minus_file,
  genome = NULL,
  keep.X = TRUE,
  keep.Y = TRUE,
  keep.M = FALSE,
  keep.nonstandard = FALSE
)
}
\arguments{
\item{genome}{Optional string for UCSC reference genome, e.g. "hg38". If
given, non-standard chromosomes are trimmed.}

\item{keep.X}{Logical indicating whether the X chromosome should be kept.}

\item{keep.Y}{Logical indicating whether the Y chromosome should be kept.}

\item{keep.M}{Logical indicating whether mitochondrial chromosomes should be
kept.}

\item{plus.file}{Path of plus strand bedGraph file.}

\item{minus.file}{Path of minus strand bedGraph file.}
}
\value{
Imports a GRanges object containing entire strand-specific reads.
  Each range is unique, and the \code{score} metadata column indicates the
  number of identical reads (which share the same 5' and 3' ends).
}
\description{
This function imports plus/minus pairs of bedGraph files. This function is
useful for when both 5'- and 3'-end information is to be maintained for each
sequenced molecule.
}
\author{
Mike DeBerardine
}
