% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import.PROseq}
\alias{import.PROseq}
\title{Import PRO-seq (or similar) bigWig files}
\usage{
import.PROseq(
  plus_file,
  minus_file,
  genome = NULL,
  keep_X = TRUE,
  keep_Y = TRUE,
  keep_M = FALSE,
  keep_nonstandard = FALSE
)
}
\arguments{
\item{genome}{Optional string for UCSC reference genome, e.g. "hg38". If
given, non-standard chromosomes are trimmed, and options for sex and
mitochondrial chromosomes are applied.}

\item{keep_X}{Logical indicating whether the X chromosome should be kept.}

\item{keep_Y}{Logical indicating whether the Y chromosome should be kept.}

\item{keep_M}{Logical indicating whether mitochondrial chromosomes should be
kept.}

\item{plus_bw}{Path of plus strand bigWig file.}

\item{minus_bw}{Path of minus strand bigWig file.}
}
\value{
Imports a GRanges object containing base-pair resolution data, with
  the \code{score} metadata column indicating readcounts at each base. All
  ranges are of width = 1.
}
\description{
This function imports plus/minus pairs of bigWig files containing
basepair-resolution data, e.g. PRO-seq or PRO-cap data.
}
\author{
Mike DeBerardine
}
