% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{importGenesUCSC}
\alias{importGenesUCSC}
\title{Import genes from UCSC}
\usage{
importGenesUCSC(
  genome,
  keep_X = TRUE,
  keep_Y = TRUE,
  keep_M = FALSE,
  keep_nonstandard = FALSE
)
}
\arguments{
\item{genome}{A string indicating the UCSC reference genome, e.g. "hg38".}

\item{keep_X}{A logical indicating if X chromosome transcripts should be
kept.}

\item{keep_Y}{A logical indicating if Y chromosome transcripts should be
kept.}
}
\value{
A GRanges object, including metadata for unique identifiers.
}
\description{
Imports all annotated genes by calling
\code{\link[GenomicFeatures:genes]{GenomicFeatures::genes}}, which provides a
single range for each annotated gene. Currently supports hg38, hg19, mm10,
mm9, dm6, and dm3. This script filters non-standard and mitochondrial
chromosomes.
}
\seealso{
\code{\link{md.import.txs.ucsc}}
}
