% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{importTxsUCSC}
\alias{importTxsUCSC}
\title{Import transcripts from UCSC}
\usage{
importTxsUCSC(
  genome,
  keep_X = TRUE,
  keep_Y = TRUE,
  keep_M = FALSE,
  keep_nonstandard = FALSE
)
}
\arguments{
\item{genome}{A string indicating the UCSC reference genome, e.g. "hg38".}

\item{keep_X}{A logical indicating if X chromosome transcripts should be
kept.}

\item{keep_Y}{A logical indicating if Y chromosome transcripts should be
kept.}
}
\value{
A GRanges object, including metadata for unique identifiers.
}
\description{

}
\seealso{
\code{\link{md.import.genes.ucsc}}
}
