% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{makeGRangesBPres}
\alias{makeGRangesBPres}
\title{Make base-pair resolution GRanges object}
\usage{
makeGRangesBPres(gr)
}
\arguments{
\item{gr}{A disjoint GRanges object.}
}
\description{
Splits up all ranges in \code{gr} to be each 1 basepair wide. All information
is preserved, including all metadata. To wit, \code{length(output.gr) =
sum(width(dataset.gr))}.
}
