% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{mergeGRangesData}
\alias{mergeGRangesData}
\title{Merge basepair-resolution GRanges objects}
\usage{
mergeGRangesData(..., field = "score", ncores = detectCores())
}
\arguments{
\item{...}{Any number of GRanges objects in which signal (e.g. readcounts)
are contained within metadata. GRanges not single-width will be coerced
using \code{\link[BRGenomics:makeGRangesBRG]{makeGRangesBRG}}.}

\item{field}{One or more metadata fields to be combined, typically the
"score" field. Fields typically contain coverage information.}

\item{ncores}{More than one core can be used to coerce non-single-width
GRanges objects using \code{makeGRangesBRG}.}
}
\value{
A disjoint, basepair-resolution (single-width) GRanges object
  comprised of all ranges found in the input GRanges objects, and whose total
  readcount is the sum of all input GRanges.
}
\description{
Merges 2 or more basepair-resolution (single-width) GRanges objects by
combining all of their ranges and associated signal (e.g. readcounts).
}
\examples{
data("PROseq") # load included PROseq data

#--------------------------------------------------#
# divide & recombine PROseq (no overlapping positions)
#--------------------------------------------------#

thirds <- floor( (1:3)/3 * length(PROseq) )
ps_1 <- PROseq[1:thirds[1]]
ps_2 <- PROseq[(thirds[1]+1):thirds[2]]
ps_3 <- PROseq[(thirds[2]+1):thirds[3]]

# re-merge
length(PROseq)
length(ps_1)
length(mergeGRangesData(ps_1, ps_2))
length(mergeGRangesData(ps_1, ps_2, ps_3))

#--------------------------------------------------#
# combine PRO-seq with overlapping positions
#--------------------------------------------------#

gr1 <- PROseq[10:13]
gr2 <- PROseq[12:15]

PROseq[10:15]

mergeGRangesData(gr1, gr2)
}
\seealso{
\code{\link[BRGenomics:makeGRangesBRG]{makeGRangesBRG}}
}
\author{
Mike DeBerardine
}
