% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_counts.R
\name{metaSubsample}
\alias{metaSubsample}
\title{Iterative Subsampling for Metaplotting}
\usage{
metaSubsample(
  dataset.gr,
  regions.gr,
  binsize = 1,
  first_output_xval = 1,
  sample_name = deparse(substitute(dataset.gr)),
  n_iter = 1000,
  prop_subsample = 0.1,
  lower = 0.125,
  upper = 0.875,
  NF = 1,
  field = "score",
  remove_empty = FALSE,
  ncores = 1
)
}
\arguments{
\item{dataset.gr}{A GRanges object in which signal is contained in metadata
(typically in the \code{"score"} field).}

\item{regions.gr}{A GRanges object containing intervals over which to
metaplot. All ranges must have the same width.}

\item{binsize}{The size of bin (number of columns, e.g. basepairs) to use for
metaplotting. Especially important for metaplots over large/sparse regions.}

\item{first_output_xval}{The relative start position of the first bin, e.g.
if regions.gr begins at 50 bases upstream of the TSS, set
\code{first_output_xval = -50}. This number only affects the x-values that
are returned, which are provided as a convenience.}

\item{sample_name}{Defaults to the name of \code{dataset.gr}. This is
included in the output as a convenience for row-binding outputs from
different samples.}

\item{n_iter}{Number of random subsampling iterations to perform. Default is
\code{1000}.}

\item{prop_subsample}{The proportion of the ranges in \code{regions.gr} (e.g.
the proportion of genes) to subsample in each iteration. The default is
\code{0.1} (10 percent).}

\item{lower}{The lower quantile of subsampled signal means to return. The
default is \code{0.125} (12.5th percentile).}

\item{upper}{The upper quantile of subsampled signal means to return. The
default is \code{0.875} (85.5th percentile).}

\item{NF}{Optional normalization factor by which to multiply the counts.}

\item{field}{The metadata field of \code{dataset.gr} to be counted.}

\item{remove_empty}{A logical indicating whether regions without signal
should be removed from the analysis.}

\item{ncores}{Number of cores to use for parallel computation. No parallel
processing is used by default, as there's no performance benefit for
typical usage with short computation times.}
}
\value{
Dataframe containing x-values, means, lower quantiles, upper
  quantiles, and the sample name (as a convenience for row-binding multiple
  of these dataframes).
}
\description{
This function performs bootstrap subsampling of mean readcounts at different
positions within regions of interest. Mean signal counts can be estimated at
base-pair resolution, or smoothed over larger bins.
}
\seealso{
\code{\link[BRGenomics:metaSubsampleMatrix]{metaSubsampleMatrix}},
  \code{\link[BRGenomics:getCountsByPositions]{getCountsByPositions}}
}
\author{
Mike DeBerardine
}
