% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_counts.R
\name{metaSubsampleScaled}
\alias{metaSubsampleScaled}
\title{Iterative Subsampling for Metaplotting (With Scaled Regions)}
\usage{
metaSubsampleScaled(
  dataset.gr,
  ...,
  linear_regions_start.gr = NULL,
  linear_regions_end.gr = NULL,
  nbins_scaled = 100,
  nbins_linear_start = unique(width(linear_regions_start.gr)),
  nbins_linear_end = unique(width(linear_regions_end.gr)),
  sample_name = deparse(substitute(dataset.gr)),
  n_iter = 1000,
  prop_subsample = 0.1,
  lower = 0.125,
  upper = 0.875,
  NF = 1,
  field = "score",
  remove_empty = FALSE,
  ncores = 1
)
}
\arguments{
\item{dataset.gr}{A GRanges object in which signal is contained in metadata
(typically in the "score" field).}

\item{...}{0 or more GRanges objects containing regions of interest over
which to do length-scaled signal counting and metaplot subsampling. The
output x-positions will be determined by the order in which these regions
are supplied, the number of bins used for counting signal within variable
length regions (\code{nbins_scaled}), and whether or not a
\code{linear_regions_start.gr} object is given.}

\item{linear_regions_start.gr}{Optional GRanges object containing regions of
interest over which to do linear (un-scaled) signal counting and metaplot
subsampling. Because no length-scaling is performed, all ranges must have
the same width. These regions will be put before any supplied regions for
length-scaled metaplot subsampling, i.e. the first
\code{nbins_linear_start} x-values will be from subsampling
\code{linear_regions_start.gr}.}

\item{linear_regions_end.gr}{Optional GRanges object containing regions of
interest over which to do linear (un-scaled) signal counting and metaplot
subsampling. Because no length-scaling is performed, all ranges must have
the same width. These regions will be placed after any supplied regions for
length-scaled metaplot subsampling, i.e. the last \code{nbins_linear_end}
x-values will be from subsampling \code{linear_regions_end.gr}.}

\item{nbins_scaled}{The number of bins to use for length scaling signal
counts.}

\item{nbins_linear_start}{The number of bins to use for counting signal
within \code{linear_regions_start.gr}. Defaults to the width of the
regions, i.e. a binsize of 1 (no binning).}

\item{nbins_linear_end}{The number of bins to use for counting signal within
\code{linear_regions_end.gr}. Defaults to the width of the regions, i.e. a
binsize of 1 (no binning).}

\item{sample_name}{Defaults to the name of \code{dataset.gr}.}

\item{n_iter}{Number of random subsampling iterations to perform. Default is
1000.}

\item{prop_subsample}{The proportion of the genelist (regions.gr) to
subsample in each iteration. The default is 0.1.}

\item{lower}{The lower quantile of subsampled signal means to return. The
default is 0.125 (12.5th percentile).}

\item{upper}{The upper quantile of subsampled signal means to return. The
default is 0.875 (85.5th percentile).}

\item{NF}{Optional normalization factor by which to multiply the counts.}

\item{field}{The metadata field of \code{dataset.gr} to be counted.}

\item{remove_empty}{A logical indicating whether regions without signal
should be removed from the analysis.}

\item{ncores}{Number of cores to use for parallel computation. As of writing,
parallel processing doesn't show any benefit for short computation times
(e.g. <1 minute for our typical experience on a laptop).}
}
\value{
Dataframe containing x-values, means, lower quantiles, upper
  quantiles, and the sample name (as a convenience for row-binding multiple
  output dataframes). X-values correspond to bins based on the input regions
  given and the specified binsizes to use.
}
\description{
This function can perform iterative metaplot subsampling in several
configurations. All arguments for regions-of-interest for metaplot
subsampling are optional, as long as at least one region is supplied. Unnamed
arguments are regions-of-interest of variable "widths" (i.e. "lengths" in
basepairs) over which to perform length-scaled metaplot subsampling.
Length-scaled metaplot subsampling involves dividing each range (e.g. each
region-of-interest) into \code{nbins_scaled} number of equally-sized bins,
and obtaining signal counts in each bin, divided by the size of the bin for
that particular region-of-interest. Non-length-scaled subsampling (as would
be done using \code{metaSubsample}) is performed on the named arguments
\code{linear_regions_start.gr} and \code{linear_regions_end.gr}. The output
is constructed in the order \code{linear_regions_start.gr}, unnamed scaled
regions (in order given), and then \code{linear_regions_end.gr}, with
x-values corresponding to the bin number.
}
\details{
The user must be able to determine the correct meaning of the bin numbers in
the final output, and for that reason arguments for binning are always
explicitly the number of bins, and not the size of the bins (as would be
possible for linear (un-scaled) regions). For example, if the user provides
\code{linear_regions_start.gr}, one unnamed GRanges for length-scaled
subsampling, and \code{linear_regions_end.gr}, the output x-values
\code{1:nbins_linear_start} will correspond to equally-sized bins in
\code{linear_regions_start.gr}; the subsequent \code{nbins_scaled} x-values
will correspond to variably-sized bins in the unnamed GRanges object given,
and the final \code{nbins_linear_end} x-values will correspond to
equally-sized bins in \code{linear_regions_end.gr}.
}
\examples{
metaSubsampleScaled(my_proseq_data, genes.early_genebodies,
genes.late_genebodies,
linear_regions_start.gr = genes.promoter_proximal,
linear_regions_end.gr = genes.cps_proximal,
nbins_scaled = 500, nbins_linear_end = 500)
}
\author{
Mike DeBerardine
}
