% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROseq-data}
\alias{PROseq-data}
\alias{PROseq}
\alias{PROseq_paired}
\title{PRO-seq data from Drosophila S2 cells}
\format{A disjoint GRanges object with 47533 ranges with 1 metadata column:
\describe{
  \item{score}{coverage of PRO-seq read 3'-ends}
  ...
}}
\source{
GEO Accession GSM1032758, run SRR611828.
}
\usage{
data(PROseq)

data(PROseq_paired)
}
\description{
PRO-seq data from chromosome 4 of Drosophila S2 cells.
}
\references{
Hojoong Kwak, Nicholas J. Fuda, Leighton J. Core, John T. Lis
  (2013). Precise Maps of RNA Polymerase Reveal How Promoters Direct
  Initiation and Pausing. \emph{Science} 339(6122): 950–953.
  \url{https://doi.org/10.1126/science.1229386}
}
\keyword{datasets}
