% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PROseq_paired}
\alias{PROseq_paired}
\title{Paired PRO-seq data from Drosophila S2 cells}
\format{A GRanges object with 52464 ranges with 1 metadata column:
\describe{
  \item{score}{number of reads sharing the same mapped 5' and 3' ends}
  ...
}}
\source{
GEO Accession GSM1032758, run SRR611828.
}
\usage{
data(PROseq_paired)
}
\description{
PRO-seq data of Drosophila S2 cells, chromosome 4. Entire mapped reads kept.
}
\details{
Hojoong Kwak, Nicholas J. Fuda, Leighton J. Core, John T. Lis (2013).
Precise Maps of RNA Polymerase Reveal How Promoters Direct Initiation and
Pausing. Science, 339(6122), 950–953.
\url{https://doi.org/10.1126/science.1229386}
}
\keyword{datasets}
