% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{subsampleBySpikeIn}
\alias{subsampleBySpikeIn}
\title{Randomly subsample reads according to spike-in normalization}
\usage{
subsampleBySpikeIn(
  dataset.gr,
  si_pattern = NULL,
  si_names = NULL,
  ctrl_pattern = NULL,
  ctrl_names = NULL,
  batch_norm = TRUE,
  RPM_units = FALSE,
  field = "score",
  sample_names = NULL,
  ncores = detectCores()
)
}
\arguments{
\item{dataset.gr, si_pattern, si_names, ctrl_pattern, ctrl_names, batch_norm, field, sample_names, ncores}{See \code{\link[BRGenomics:getSpikeInNFs]{getSpikeInNFs}}}

\item{RPM_units}{If set to \code{TRUE}, the final readcount values will be
converted to \code{negative control RPM}.}
}
\value{
A list of GRanges parallel to \code{dataset.gr}, but containing fewer
  reads.
}
\description{
Randomly subsample reads according to spike-in normalization
}
\seealso{
\code{\link[BRGenomics:getSpikeInCounts]{getSpikeInCounts}},
  \code{\link[BRGenomics:getSpikeInNFs]{getSpikeInNFs}}
}
\author{
Mike DeBerardine
}
