% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{subsampleGRanges}
\alias{subsampleGRanges}
\title{Randomly subsample reads from GRanges dataset}
\usage{
subsampleGRanges(dataset.gr, n = NULL, prop = NULL, field = "score")
}
\arguments{
\item{dataset.gr}{A GRanges object in which signal (e.g. readcounts) are
contained within metadata.}

\item{n}{Number of reads to subsample. Either \code{n} or \code{prop} can be
given.}

\item{prop}{Proportion of total signal to subsample.}

\item{field}{The metadata field of \code{dataset.gr} that contains readcounts
for reach position. If each range represents a single read, set \code{field
= NULL}}
}
\value{
A GRanges object identical in format to \code{dataset.gr}, but
  containing a random subset of its data. If \code{field != NULL}, the length
  of the output cannot be known \emph{a priori}, but the sum of its score
  can.
}
\description{
Random subsampling is not performed on ranges, but on reads. Readcounts
should be given as a metadata field (usually "score"). This function can also
subsample ranges directly if \code{field = NULL}, but the \code{sample}
function can be used in this scenario.
}
\section{Use with normalized readcounts}{
 If the metadata field contains
  normalized readcounts, an attempt will be made to infer the normalization
  factor based on the lowest signal value found in the specified field.
}

\examples{
data("PROseq") # load included PROseq data

#--------------------------------------------------#
# sample 10\% of the reads of a GRanges with signal coverage
#--------------------------------------------------#

ps_sample <- subsampleGRanges(PROseq, prop = 0.1)

# cannot predict number of ranges (positions) that will be sampled
length(PROseq)
length(ps_sample)

# 1/10th the score is sampled
sum(score(PROseq))
sum(score(ps_sample))

#--------------------------------------------------#
# Sample 10\% of ranges (e.g. if each range represents one read)
#--------------------------------------------------#

ps_sample <- subsampleGRanges(PROseq, prop = 0.1, field = NULL)

length(PROseq)
length(ps_sample)

# Alternatively
ps_sample <- sample(PROseq, 0.1 * length(PROseq))
length(ps_sample)
}
\author{
Mike DeBerardine
}
