% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_functions.R
\name{subsetRegionsBySignal}
\alias{subsetRegionsBySignal}
\title{Subset regions of interest by highest signal}
\usage{
subsetRegionsBySignal(
  regions.gr,
  dataset.gr,
  regions_quantile,
  field = "score",
  order_by_rank = FALSE,
  density = FALSE
)
}
\arguments{
\item{regions.gr}{A GRanges object containing regions of interest.}

\item{dataset.gr}{A GRanges object in which signal is contained in metadata
(typically in the "score" field).}

\item{regions_quantile}{The proportion of regions.gr to return, e.g. if
\code{regions_quantile = 0.2}, the top 20\% of regions by signal are
returned.}

\item{field}{The metadata field of \code{dataset.gr} to be counted.}

\item{order_by_rank}{Logical indicating if genes should be returned in order
of their expression. If \code{FALSE} (the default), genes are sorted by
their positions.}

\item{density}{A logical indicating whether signal counts should be
normalized to the width of ranges in \code{regions.gr}. By default, the
function only considers the total signal in each range.}
}
\value{
A GRanges object of length \code{length(regions.gr) *
  regions_quantile}.
}
\description{
Subsets regions based on signal in a dataset, taking only the top quantile of
regions.
}
