% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{tidyChromosomes}
\alias{tidyChromosomes}
\title{Remove odd chromosomes from GRanges objects}
\usage{
tidyChromosomes(
  gr,
  keep_X = TRUE,
  keep_Y = TRUE,
  keep_M = FALSE,
  keep_nonstandard = FALSE
)
}
\arguments{
\item{gr}{Any GRanges object, however the object should have a standard
genome set, e.g. \code{genome(gr) <- "hg38"}}

\item{keep_X, keep_Y, keep_M, keep_nonstandard}{Logicals indicating which
non-autosomes should be kept. By default, sex chromosomes are kept, but
mitochondrial and non-standard chromosomes are removed.}
}
\description{
This convenience function removes non-standard, mitochondrial, and/or sex
chromosomes from any GRanges object. For the chromosomes being removed, any
ranges found on those chromosomes are removed, and the chromosomes are also
removed from \code{seqinfo}. Standard chromosomes are defined using the
\code{\link[GenomeInfoDb:standardChromosomes]{standardChromosomes}} function
from the \code{GenomeInfoDb} package.
}
\seealso{
\code{\link[GenomeInfoDb:standardChromosomes]{
   GenomeInfoDb::standardChromosomes}},
}
\author{
Mike DeBerardine
}
