% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndimensional_binning.R
\name{binNdimensions}
\alias{binNdimensions}
\title{N-dimensional binning of data by quantiles}
\usage{
binNdimensions(..., quantiles = 10)
}
\arguments{
\item{...}{A single dataframe, or any number of lists or
vectors containing different measurements across the same samples. If lists
or vectors are given, they must all have the same lengths.}

\item{quantiles}{Either a number giving the number of quantiles to use for
all dimensions (default = 10), or a vector containing the number of
quantiles to use for each dimension of input data given.}
}
\value{
A dataframe containing indices in \code{1:quantiles} for each
  datapoint in each dimension.
}
\description{
This function takes in data along 1 or more dimensions, and for each
dimension evenly divides the data in evenly-sized quantiles from the minimum
value to the maximum value. For each input data point, indices are returned
giving the bin in each dimension.
}
\author{
Mike DeBerardine
}
