% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRGenomics-package.R
\docType{package}
\name{BRGenomics-package}
\alias{BRGenomics-package}
\alias{BRGenomics}
\title{BRGenomics: Tools for the Efficient Analysis of High-Resolution Genomics Data}
\description{
BRGenomics provides useful functions for analyzing genomics data
  at base-pair resolution, and for doing so in a way that maximizes
  compatibility with the wide array of packages available through
  Bioconductor.

  For interactive documentation with code examples, see the online
  documentation: \url{https://mdeber.github.io/}
}
\author{
Mike DeBerardine \email{mike.deberardine@gmail.com}
}
