% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_functions.R
\name{genebodies}
\alias{genebodies}
\title{Extract Genebodies}
\usage{
genebodies(
  genelist,
  start = 300L,
  end = -300L,
  fix.start = "start",
  fix.end = "end",
  min.window = 0L
)
}
\arguments{
\item{genelist}{A GRanges object containing genes of interest.}

\item{start}{Depending on \code{fix.start}, the distance from either the
strand-specific start or end site to begin the returned ranges. If
positive, the returned range will begin downstream of the reference
position; negative numbers are used to return sites upstream of the
reference. Set \code{start = 0} to return the reference position.}

\item{end}{Identical to the \code{start} argument, but defines the
strand-specific end position of returned ranges. \code{end} must be
downstream of \code{start}.}

\item{fix.start}{The reference point to use for defining the strand-specific
start positions of returned ranges, either \code{"start"} or \code{"end"}.}

\item{fix.end}{The reference point to use for defining the strand-specific
end positions of returned ranges, either \code{"start"} or \code{"end"}.
Cannot be set to \code{"start"} if \code{fix.start = "end"}.}

\item{min.window}{When \code{fix.start = "start"} and \code{fix.end = "end"},
\code{min.window} defines the minimum size (width) of a returned range.
However, when \code{fix.end = fix.start}, all returned ranges have the same
width, and \code{min.window} simply size-filters the input ranges.}
}
\value{
A GRanges object that may be shorter than \code{genelist} due to
  filtering of short ranges. For example, using the default arguments, genes
  shorter than 600 bp would be removed.
}
\description{
This function returns ranges that are defined relative to the strand-specific
start and end sites of regions of interest (usually genes).
}
\details{
Unlike
  \code{\link[GenomicRanges:intra-range-methods]{GenomicRanges::promoters}},
  distances can be defined to be upstream or downstream by changing the sign
  of the argument, and both the start and end of the returned regions can be
  defined in terms of the strand-specific start or end site of the input
  ranges. For example, \code{genebodies(txs, -50, 150, fix.end = "start")} is
  equivalent to \code{promoters(txs, 50, 151)} (the downstream edge is off by
  1 because \code{promoters} keeps the downstream interval closed). The
  default arguments return ranges that begin 300 bases downstream of the
  original start positions, and end 300 bases upstream of the original end
  positions.
}
\examples{
data("txs_dm6_chr4") # load included transcript data
txs <- txs_dm6_chr4[c(1, 2, 167, 168)]

txs

#--------------------------------------------------#
# genebody regions from 300 bp after the TSS to
# 300 bp before the polyA site
#--------------------------------------------------#

genebodies(txs, 300, -300)

#--------------------------------------------------#
# promoter-proximal region from 50 bp upstream of
# the TSS to 100 bp downstream of the TSS
#--------------------------------------------------#

promoters(txs, 50, 101)

genebodies(txs, -50, 100, fix.end = "start")

#--------------------------------------------------#
# region from 500 to 1000 bp after the polyA site
#--------------------------------------------------#

genebodies(txs, 500, 1000, fix.start = "end")
}
\seealso{
\code{\link[GenomicRanges:intra-range-methods]{intra-range-methods}}
}
\author{
Mike DeBerardine
}
