% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_functions.R
\name{genebodies}
\alias{genebodies}
\title{Extract Genebodies}
\usage{
genebodies(
  genelist,
  start = 300,
  end = -300,
  fix_start = "start",
  fix_end = "end",
  min_window = 500
)
}
\arguments{
\item{genelist}{A GRanges object containing genes of interest.}

\item{start}{When \code{fix = "start"}, the distance downstream of the TSS
where the new ranges should begin.}

\item{end}{Where the ranges should end. When \code{end = 0}, the returned
ranges keep the original end sites. When \code{end < 0}, the new ranges
will end \code{abs(end)} number of bases before the original end site. When
\code{end > 0} and \code{fix = "start"}, the new ends are fixed to
\code{end} number of bases from the \emph{original} start site.}

\item{min.window}{The minimum size of a returned genebody length, after
accounting for \code{start} and \code{end} parameters.}

\item{fix}{If set to "end", function will return ranges centered around the
end of the ranges. Negative values for \code{start} will begin the output
ranges a fixed distance before the end of the input ranges; positive values
will begin the ranges after the ends of the input ranges. The behavior of
\code{end} is the same as for \code{start}, and errors will be returned if
\code{end < start}.}
}
\value{
A GRanges object that may be shorter than \code{genelist} due to loss
  of short ranges.
}
\description{
Analagous to \code{\link[GenomicRanges:promoters]{GenomicRanges::promoters}},
this function returns ranges that start and end downstream of the TSS. When
\code{fix = "start"}, the function behaves differently depending on the sign
of the \code{end} parameter. Currently, there's no way to set the ranges to
continue a fixed distance past the original end sites when \code{fix =
"start"}.
}
