% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_counting.R
\name{getCountsByPositions}
\alias{getCountsByPositions}
\title{Get signal count matrix within regions of interest}
\usage{
getCountsByPositions(
  dataset.gr,
  regions.gr,
  binsize = 1,
  field = "score",
  remove_empty = FALSE
)
}
\arguments{
\item{dataset.gr}{A GRanges object in which signal is contained in metadata
(typically in the "score" field).}

\item{regions.gr}{A GRanges object containing all the regions of interest.
All ranges must have the same width!}

\item{binsize}{Size of bins (in bp) to use for counting within each range of
\code{regions.gr}. Note that counts will \emph{not} be length-normalized.}

\item{field}{The metadata field of \code{dataset.gr} to be counted.}

\item{remove_empty}{Logical indicating whether regions without signal should
be removed.}
}
\value{
A matrix containing a row for each range in regions.gr, and a column
  for each bin.
}
\description{
Get signal count matrix within regions of interest
}
