% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_counting.R
\name{getCountsByRegions}
\alias{getCountsByRegions}
\title{Signal counts in regions of interest}
\usage{
getCountsByRegions(dataset.gr, regions.gr, field = "score")
}
\arguments{
\item{dataset.gr}{A GRanges object in which signal is contained in metadata
(typically in the "score" field).}

\item{regions.gr}{A GRanges object containing all the regions of interest.}

\item{field}{The metadata field of \code{dataset.gr} to be counted.}
}
\value{
Returns a vector the same length as \code{regions.gr} containing
  signal found in each range.
}
\description{
Count signal (e.g. read coverage) of data in each region of interest.
}
