% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_counting.R
\name{getPausingIndices}
\alias{getPausingIndices}
\title{Calculate pausing indices from user-supplied promoters & genebodies}
\usage{
getPausingIndices(
  dataset.gr,
  promoters.gr,
  genebodies.gr,
  field = "score",
  length_normalize = TRUE,
  remove_empty = FALSE
)
}
\arguments{
\item{dataset.gr}{A GRanges object in which signal is contained in metadata
(typically in the "score" field).}

\item{promoters.gr}{A GRanges object containing all the regions of interest.
The sum of all signal counts within is the pause index numerator.}

\item{genebodies.gr}{A GRanges object containing all the regions of interest.
The sum of all signal counts within is the pause index denominator.}

\item{field}{The metadata field of \code{dataset.gr} to be counted, i.e. that
contains the readcounts of interest.}

\item{length_normalize}{A logical indicating if signal counts within regions
of interest should be length normalized. The default is TRUE, which is
recommended, especially if input regions don't all have the same width.}

\item{remove_empty}{A logical indicating if genes without any signal should
be removed. The default is FALSE.}
}
\value{
A vector of length given by the length of the genelist (or possibly
  shorter if \code{remove_empty = TRUE}).
}
\description{
Pausing index (PI) is calculated for each gene (within matched
\code{promoters.gr} and \code{genebodies.gr}) as promoter signal counts
divided by genebody signal counts. If \code{length.normalize = TRUE}
(recommended), the signal counts within each range in \code{promoters.gr} and
\code{genebodies.gr} are divided by their respective range widths (region
lengths) before pausing indices are calculated.
}
