% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{getStrandedCoverage}
\alias{getStrandedCoverage}
\title{Get strand-specific coverage}
\usage{
getStrandedCoverage(gr, field = "score")
}
\arguments{
\item{field}{The name of the field that contains readcounts. If no metadata
field contains readcounts, and each range represents a single read, set to
NULL.}

\item{dataset.gr}{A GRanges object either containing ranges for each read, or
one in which readcounts for individual ranges are contained in metadata
(typically in the "score" field).}
}
\description{
Computes strand-specific coverage signal, and returns a GRanges object with
signal in the "score" metadata column. Function also works for
non-strand-specific data. Note that output is not automatically converted
into a "basepair-resolution" GRanges object.
}
\author{
Mike DeBerardine
}
