% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import-functions}
\alias{import-functions}
\alias{import_bigWig}
\alias{import_bedGraph}
\title{Import basepair-resolution files}
\usage{
import_bigWig(
  plus_file,
  minus_file,
  genome = NULL,
  keep.X = TRUE,
  keep.Y = TRUE,
  keep.M = FALSE,
  keep.nonstandard = FALSE
)

import_bedGraph(
  plus_file,
  minus_file,
  genome = NULL,
  keep.X = TRUE,
  keep.Y = TRUE,
  keep.M = FALSE,
  keep.nonstandard = FALSE
)
}
\arguments{
\item{plus_file, minus_file}{Paths for strand-specific input files.}

\item{genome}{Optional string for UCSC reference genome, e.g. "hg38". If
given, non-standard chromosomes are trimmed, and options for sex and
mitochondrial chromosomes are applied.}

\item{keep.X, keep.Y, keep.M, keep.nonstandard}{Logicals indicating which
non-autosomes should be kept. By default, sex chromosomes are kept, but
mitochondrial and non-standard chromosomes are removed.}
}
\description{
Import basepair-resolution files
}
\details{
Imports a GRanges object containing base-pair resolution data, with
  the \code{score} metadata column indicating the number of reads represented
  by each range.

  \code{import_bedGraph} is useful for when both 5'- and 3'-end information
  is to be maintained for each sequenced molecule. It effectively imports the
  entire read.

  For \code{import_bigWig}, all ranges are of width = 1.
}
\examples{
#--------------------------------------------------#
# Import PRO-seq bigWigs -> coverage of 3' bases
#--------------------------------------------------#

# get local address for included bigWig files
p.bw <- system.file("extdata", "PROseq_dm6_chr4_plus.bw",
                    package = "BRGenomics")
m.bw <- system.file("extdata", "PROseq_dm6_chr4_minus.bw",
                    package = "BRGenomics")

# import bigWigs
PROseq <- import_bigWig(p.bw, m.bw, genome = "dm6")
PROseq

#--------------------------------------------------#
# Import PRO-seq bedGraphs -> whole reads (matched 5' and 3' ends)
#--------------------------------------------------#

# get local address for included bedGraph files
p.bg <- system.file("extdata", "PROseq_dm6_chr4_plus.bedGraph",
                    package = "BRGenomics")
m.bg <- system.file("extdata", "PROseq_dm6_chr4_minus.bedGraph",
                    package = "BRGenomics")

# import bedGraphs
PROseq_paired <- import_bedGraph(p.bg, m.bg, genome = "dm6")
PROseq_paired
}
\seealso{
\code{\link[BRGenomics:tidyChromosomes]{tidyChromosomes}},
  \code{\link[rtracklayer:import]{rtracklayer::import}}
}
\author{
Mike DeBerardine
}
