% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_bam}
\alias{import_bam}
\title{Import bam files}
\usage{
import_bam(
  file,
  mapq = 20,
  revcomp = FALSE,
  trim.to = c("whole", "5p", "3p", "center"),
  ignore.strand = FALSE,
  field = "score"
)
}
\arguments{
\item{file}{Path of a bam file.}

\item{mapq}{Filter reads by a minimum MAPQ score. This is the correct way to
filter multi-aligners.}

\item{revcomp}{Logical indicating if aligned reads should be
reverse-complemented.}

\item{trim.to}{Option for selecting specific bases from the reads, applied
after the \code{revcomp} option. By default, the entire read is maintained.
Other options are to take only the 5' base, only the 3' base, or the only
the center base of the read.}

\item{ignore.strand}{Logical indicating if the strand information should be
discarded. If \code{TRUE}, strand information is discarded \emph{after}
\code{revcomp} and \code{trim.to} options are applied.}

\item{field}{Metadata field name to use for readcounts, usually "score". If
set to \code{NULL}, identical reads (or identical positions if
\code{trim.to} options applied) are not combined, and the length of the
output GRanges will be equal to the number of input reads.}
}
\value{
A GRanges object.
}
\description{
CURRENTLY ONLY FOR SINGLE-END SEQUENCING. Patch for paired-end support is coming soon.
}
\examples{
# get local address for included bam file
ps.bam <- system.file("extdata", "PROseq_dm6_chr4.bam",
                      package = "BRGenomics")

#--------------------------------------------------#
# Import entire reads
#--------------------------------------------------#

# Note that PRO-seq reads are sequenced as reverse complement
import_bam(ps.bam, revcomp = TRUE)

#--------------------------------------------------#
# Import entire reads, 1 range per read
#--------------------------------------------------#

import_bam(ps.bam, revcomp = TRUE, field = NULL)

#--------------------------------------------------#
# Import PRO-seq reads at basepair-resolution
#--------------------------------------------------#

# the typical manner to import PRO-seq data:
import_bam(ps.bam, revcomp = TRUE, trim.to = "3p")

#--------------------------------------------------#
# Import 5' ends of PRO-seq reads
#--------------------------------------------------#

# will include bona fide TSSes as well as hydrolysis products
import_bam(ps.bam, revcomp = TRUE, trim.to = "5p")
}
