% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{import_bam}
\alias{import_bam}
\alias{import_bam_PROseq}
\alias{import_bam_PROcap}
\alias{import_bam_ATACseq}
\title{Import bam files}
\usage{
import_bam(
  file,
  mapq = 20,
  revcomp = FALSE,
  shift = 0L,
  trim.to = c("whole", "5p", "3p", "center"),
  ignore.strand = FALSE,
  field = "score",
  paired_end = NULL,
  yieldSize = 250000
)

import_bam_PROseq(
  file,
  mapq = 20,
  revcomp = TRUE,
  shift = -1L,
  trim.to = "3p",
  ignore.strand = FALSE,
  field = "score",
  paired_end = NULL,
  yieldSize = 250000
)

import_bam_PROcap(
  file,
  mapq = 20,
  revcomp = FALSE,
  shift = 0L,
  trim.to = "5p",
  ignore.strand = FALSE,
  field = "score",
  paired_end = NULL,
  yieldSize = 250000
)

import_bam_ATACseq(
  file,
  mapq = 20,
  revcomp = FALSE,
  shift = c(4, -5),
  trim.to = "whole",
  ignore.strand = TRUE,
  field = "score",
  paired_end = TRUE,
  yieldSize = 250000
)
}
\arguments{
\item{file}{Path of a bam file.}

\item{mapq}{Filter reads by a minimum MAPQ score. This is the correct way to
filter multi-aligners.}

\item{revcomp}{Logical indicating if aligned reads should be
reverse-complemented.}

\item{shift}{Either an integer giving the number of bases by which to shift
the entire read upstream or downstream, or a pair of integers indicating
shifts to be applied to the 5' and 3' ends of the reads, respectively.
Shifting is strand-specific, with negative numbers shifting the reads
upstream, and positive numbers shiftem them downstream. This option is
applied \emph{after} the \code{revcomp}, but before \code{trim.to} and
\code{ignore.strand} options are applied.}

\item{trim.to}{Option for selecting specific bases from the reads, applied
after the \code{revcomp} and \code{shift} options. By default, the entire
read is maintained. Other options are to take only the 5' base, only the 3'
base, or the only the center base of the read.}

\item{ignore.strand}{Logical indicating if the strand information should be
discarded. If \code{TRUE}, strand information is discarded \emph{after}
\code{revcomp}, \code{trim.to}, and \code{shift} options are applied.}

\item{field}{Metadata field name to use for readcounts, usually "score". If
set to \code{NULL}, identical reads (or identical positions if
\code{trim.to} options applied) are not combined, and the length of the
output GRanges will be equal to the number of input reads.}

\item{paired_end}{Logical indicating if reads should be treated as paired end
reads. When set to \code{NULL} (the default), a test is performed to
determine if the bam file contains any paired-end reads.}

\item{yieldSize}{The number of bam file records to process simultaneously,
e.g. the "chunk size". Setting a higher chunk size will use more memory,
which can increase speed if there is enough memory available. If chunking
is not necessary, set to \code{NA}.}
}
\value{
A GRanges object.
}
\description{
Import single-end or paired-end bam files as GRanges objects, with various
processing options.
}
\details{
If function produces an error, make the \code{paired_end} parameter
  explicit, i.e. \code{TRUE} or \code{FALSE}.
}
\examples{
# get local address for included bam file
ps.bam <- system.file("extdata", "PROseq_dm6_chr4.bam",
                      package = "BRGenomics")

#--------------------------------------------------#
# Import entire reads
#--------------------------------------------------#

# Note that PRO-seq reads are sequenced as reverse complement
import_bam(ps.bam, revcomp = TRUE, paired_end = FALSE)

#--------------------------------------------------#
# Import entire reads, 1 range per read
#--------------------------------------------------#

import_bam(ps.bam, revcomp = TRUE, field = NULL,
           paired_end = FALSE)

#--------------------------------------------------#
# Import PRO-seq reads at basepair-resolution
#--------------------------------------------------#

# the typical manner to import PRO-seq data:
import_bam(ps.bam, revcomp = TRUE, trim.to = "3p",
           paired_end = FALSE)

#--------------------------------------------------#
# Import PRO-seq reads, removing the run-on base
#--------------------------------------------------#

# the best way to import PRO-seq data; removes the
# most 3' base, which was added in the run-on
import_bam(ps.bam, revcomp = TRUE, trim.to = "3p",
           shift = -1, paired_end = FALSE)

#--------------------------------------------------#
# Import 5' ends of PRO-seq reads
#--------------------------------------------------#

# will include bona fide TSSes as well as hydrolysis products
import_bam(ps.bam, revcomp = TRUE, trim.to = "5p",
           paired_end = FALSE)
}
\references{
Hojoong Kwak, Nicholas J. Fuda, Leighton J. Core, John T. Lis
  (2013). Precise Maps of RNA Polymerase Reveal How Promoters Direct
  Initiation and Pausing. \emph{Science} 339(6122): 950–953.
  \url{https://doi.org/10.1126/science.1229386}

  Jason D. Buenrostro, Paul G. Giresi, Lisa C. Zaba, Howard Y. Chang, William
  J. Greenleaf (2013). Transposition of native chromatin for fast and
  sensitive epigenomic profiling of open chromatin, dna-binding proteins and
  nucleosome position. \emph{Nature Methods} 10: 1213–1218.
  \url{https://doi.org/10.1038/nmeth.2688}
}
\author{
Mike DeBerardine & Nate Tippens
}
