% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{mergeGRangesData}
\alias{mergeGRangesData}
\title{Merge base-pair resolution GRanges objects}
\usage{
mergeGRangesData(..., field = "score", ncores = detectCores())
}
\arguments{
\item{...}{Any number of GRanges objects in which signal (e.g. readcounts)
are contained within metadata.}
}
\value{
A single GRange object containing all sites of the input objects, and
  the sum of all scores at all sites.
}
\description{
Merges 2 or more GRanges objects. For each object, the range widths must all
be 1, and the \code{score} metadata column contains coverage information at
each site.
}
\author{
Mike DeBerardine
}
