% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{mergeGRangesData}
\alias{mergeGRangesData}
\title{Merge base-pair resolution GRanges objects}
\usage{
mergeGRangesData(..., field = "score", ncores = detectCores())
}
\arguments{
\item{...}{One or more additional GRanges objects also fitting the above
description.}

\item{dataset.gr}{A GRanges object fitting the above description.}
}
\value{
A single GRange object containing all sites of the input objects, and
  the sum of all scores at all sites.
}
\description{
Merges 2 or more GRanges objects. For each object, the range widths must all
be 1, and the \code{score} metadata column contains coverage information at
each site.
}
\seealso{
\code{\link{md.import.bigWigs}}
}
\author{
Mike DeBerardine
}
