% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_counts.R
\name{metaSubsampleMatrix}
\alias{metaSubsampleMatrix}
\title{Iterative Subsampling for Metaplotting (On Count Matrices)}
\usage{
metaSubsampleMatrix(
  counts.mat,
  binsize = 1,
  first_output_xval = 1,
  sample_name = deparse(substitute(dataset.gr)),
  n_iter = 1000,
  prop_subsample = 0.1,
  lower = 0.125,
  upper = 0.875,
  NF = 1,
  ncores = 1
)
}
\arguments{
\item{counts.mat}{A matrix of signal counts in which rows are regions of
interest and columns are sites/bins in each region.}

\item{binsize}{The size of bin (number of columns, e.g. basepairs) to use for
metaplotting. Especially important for metaplots over large/sparse regions.}

\item{first_output_xval}{The relative start position of the first bin, e.g.
if regions.gr begins at 50 bases upstream of the TSS, set
\code{first_output_xval = -50}. This number only affects the x-values that
are returned, which are provided as a convenience.}

\item{sample_name}{Defaults to the name of \code{dataset.gr}.}

\item{n_iter}{Number of random subsampling iterations to perform. Default is
1000.}

\item{prop_subsample}{The proportion of rows to subsample in each iteration.
The default is 0.1.}

\item{lower}{The lower quantile of subsampled signal means to return. The
default is 0.125 (12.5th percentile).}

\item{upper}{The upper quantile of subsampled signal means to return. The
default is 0.875 (85.5th percentile).}

\item{NF}{Optional normalization factor by which to multiply the counts.}

\item{ncores}{Number of cores to use for parallel computation. As of writing,
parallel processing doesn't show any benefit for short computation times
(e.g. <1 minute for our typical experience on a laptop).}
}
\value{
Dataframe containing x-values, means, lower quantiles, upper
  quantiles, and the sample name (as a convenience for row-binding multiple
  of these dataframes).
}
\description{
In the most general sense, this function performs iterations of randomly
subsampling rows of a matrix, and returns a summary of mean values calculated
for each column. The typical application is for generating metaplots, with
the typical input being a matrix in which each row is a gene or other region
of interest, each column is a position within that gene (either a specific
basepair or a bin), and element values are signal (e.g. read counts) within
those positions.
}
\author{
Mike DeBerardine
}
