% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{subsampleGRanges}
\alias{subsampleGRanges}
\title{Randomly subsample reads from GRanges dataset}
\usage{
subsampleGRanges(dataset.gr, n = NULL, prop = NULL, field = "score")
}
\arguments{
\item{dataset.gr}{A GRanges object in which signal (e.g. readcounts) are
contained within metadata.}

\item{n}{Number of reads to subsample. Either \code{n} or \code{prop} can be
given.}

\item{prop}{Proportion of total signal to subsample.}

\item{field}{}
}
\description{
Currently only works if signal is integer
}
\author{
Mike DeBerardine
}
