% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{subsampleGRanges}
\alias{subsampleGRanges}
\title{Randomly subsample reads from GRanges dataset}
\usage{
subsampleGRanges(dataset.gr, n = NULL, prop = NULL, field = "score")
}
\arguments{
\item{dataset.gr}{A GRanges object in which signal (e.g. readcounts) are
contained within metadata.}

\item{n}{Number of reads to subsample. Either \code{n} or \code{prop} can be
given.}

\item{prop}{Proportion of total signal to subsample.}

\item{field}{The metadata field of \code{dataset.gr} that contains readcounts
for reach position. If each range represents a single read, set \code{field
= NULL}}
}
\description{
Random subsampling is not performed on ranges, but on reads. Readcounts
should be given as a metadata field (usually "score"), and should normally be
integers. If normalized readcounts are given, an attempt will be made to
infer the normalization factor based on the least-common-multiple of the
signal found in the specified field. This function can also subsample ranges
directly if \code{field = NULL}, but the \code{sample} function can be used
in this scenario.
}
\author{
Mike DeBerardine
}
