% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{tidyChromosomes}
\alias{tidyChromosomes}
\title{Remove odd chromosomes from GRanges objects}
\usage{
tidyChromosomes(
  gr,
  keep.X = TRUE,
  keep.Y = TRUE,
  keep.M = FALSE,
  keep.nonstandard = FALSE
)
}
\arguments{
\item{gr}{Any GRanges object, or any another object with associated
\code{seqinfo} (or a \code{Seqinfo} object itself). The object should
typically have a standard genome associated with it, e.g. \code{genome(gr)
<- "hg38"}.}

\item{keep.X, keep.Y, keep.M, keep.nonstandard}{Logicals indicating which
non-autosomes should be kept. By default, sex chromosomes are kept, but
mitochondrial and non-standard chromosomes are removed.}
}
\value{
A GRanges object in which both ranges and \code{seqinfo} associated
  with trimmed chromosomes have been removed.
}
\description{
This convenience function removes non-standard, mitochondrial, and/or sex
chromosomes from any GRanges object.
}
\details{
Standard chromosomes are defined using the
  \code{\link[GenomeInfoDb:standardChromosomes]{standardChromosomes}}
  function from the \code{GenomeInfoDb} package.
}
\examples{
# make a GRanges
chrom <- c("chr2", "chr3", "chrX", "chrY", "chrM", "junk")
gr <- GRanges(seqnames = chrom,
              ranges = IRanges(start = 2*(1:6), end = 3*(1:6)),
              strand = "+",
              seqinfo = Seqinfo(chrom))
genome(gr) <- "hg38"

gr

tidyChromosomes(gr)

tidyChromosomes(gr, keep.M = TRUE)

tidyChromosomes(gr, keep.M = TRUE, keep.Y = FALSE)

tidyChromosomes(gr, keep.nonstandard = TRUE)
}
\seealso{
\code{\link[GenomeInfoDb:standardChromosomes]{
GenomeInfoDb::standardChromosomes}}
}
\author{
Mike DeBerardine
}
